+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Minor Changes
 
+- Added support for passing in a null `tokenId` when using NFT Webhook Filters, which allows you to listen to all token ids in a collection.
 - Added `NftNamespace.getMintedNfts()` to fetch all the NFTs an owner address minted, optionally filtered by a set of specific NFT contracts.
 
 ## 2.3.0